fix: resolve comments and useAvailableWallets hook
