--- packages/wallets/src/components/AddMoreWalletsCarousel/index.ts
@@ -1,4 +1,0 @@
-import AddMoreWalletsCarousel from './AddMoreWalletsCarousel';
-import './AddMoreWalletsCarousel.scss';
-
-export default AddMoreWalletsCarousel;
