+++ packages/wallets/src/components/WalletsAddMoreCarousel/index.ts
@@ -1,0 +1,4 @@
+import WalletsAddMoreCarousel from './WalletsAddMoreCarousel';
+import './WalletsAddMoreCarousel.scss';
+
+export default WalletsAddMoreCarousel;
